LayoutTests/imported/w3c/web-platform-tests/css/css-inline/baseline-source/baseline-source-last-001-expected.txt




























PASS .target > * 1
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<span data-offset-y="65"></span>
offsetTop expected 65 but got 80
PASS .target > * 4
PASS .target > * 5
PASS .target > * 6
FAIL .target > * 7 assert_equals:
<span data-offset-y="65"></span>
offsetTop expected 65 but got 125
PASS .target > * 8
FAIL .target > * 9 assert_equals:
<span data-offset-y="95"></span>
offsetTop expected 95 but got 110
PASS .target > * 10
FAIL .target > * 11 assert_equals:
<span data-offset-y="65"></span>
offsetTop expected 65 but got 35
PASS .target > * 12
FAIL .target > * 13 assert_equals:
<span data-offset-y="65"></span>
offsetTop expected 65 but got 35
PASS .target > * 14
FAIL .target > * 15 assert_equals:
<span data-offset-y="90"></span>
offsetTop expected 90 but got 20
PASS .target > * 16
FAIL .target > * 17 assert_equals:
<span data-offset-y="90"></span>
offsetTop expected 90 but got 60
PASS .target > * 18
FAIL .target > * 19 assert_equals:
<span data-offset-y="135"></span>
offsetTop expected 135 but got 75
PASS .target > * 20
FAIL .target > * 21 assert_equals:
<span data-offset-y="135"></span>
offsetTop expected 135 but got 220
PASS .target > * 22
